Mejoramiento general de estilo
