+++ .gitignore
@@ -1,1 +1,2 @@
 .vscode
+compile_flags.txt
